Split pyi typing for io module
